Rollback Permalink Change
I decided to roll back the permalink option on the posts since it
will create unnecessary clutter in the generated gh-pages. However
I did decide to keep the excerpt option since it will help keep
blog posts simple. Plus it helps the reader have an idea of what
may be written in the post. I'll propogate excerpts for the rest
of the posts in a later commit.