Hardcoded Header Links
I coded in the HTML links to various parts of the website and
external links in the header. The reason for this was to have
control of the link order, and have a consistency when I customize
the header of the discourse site. Plus I fix a error in the latest
blog post that caused it not to serve on my local build.