Discourse Comments Take 2
I had to try a different approach to get comments to work. I
relized that I had to add a type for defaults in _config.yml. I
also just added the script in the post layout since posts will be
the only place we need comments. Now since the liquid text
condition works now we just need to see if discourse dose it's job.

Once again, I apoligise if this breaks our sites.